Default lists
When Draupnir is provisioned in the appservice, a policy
list is created for the user. We don't need to tell
them to create a new list to get started, only
how to create them for the future.

The concept of a default list does not exist in Draupnir,
there is only a legacy command that updates
the account data that was used by Mjolnir for the setting.
It will probably be removed.